update markers and finish revision controller
