+++ components/function-controller/config/rbac/role.yaml
@@ -9,32 +9,10 @@ rules:
 - apiGroups:
   - ""
   resources:
-  - configmaps
-  - secrets
-  - serviceaccounts
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - ""
-  resources:
   - events
   verbs:
   - create
   - patch
-- apiGroups:
-  - ""
-  resources:
-  - namespaces
-  verbs:
-  - get
-  - list
-  - watch
 - apiGroups:
   - batch
   resources:
@@ -54,6 +32,36 @@ rules:
   - jobs/status
   verbs:
   - get
+- apiGroups:
+  - ""
+  resources:
+  - configmaps
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+- apiGroups:
+  - ""
+  resources:
+  - namespaces
+  verbs:
+  - get
+  - list
+  - watch
+- apiGroups:
+  - ""
+  resources:
+  - secrets
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
 - apiGroups:
   - serverless.kyma-project.io
   resources:
@@ -78,11 +86,23 @@ rules:
   - serving.knative.dev
   resources:
   - revisions
+  verbs:
+  - deletecollection
+  - list
+- apiGroups:
+  - serving.knative.dev
+  resources:
+  - revisions
+  - services
+  verbs:
+  - get
+- apiGroups:
+  - serving.knative.dev
+  resources:
   - services
   verbs:
   - create
   - delete
-  - deletecollection
   - get
   - list
   - patch
