+++ components/function-controller/internal/controllers/kubernetes/kubernetes_controller.go
@@ -55,6 +55,7 @@ func (r *Reconciler) SetupWithManager(mgr ctrl.Manager) error {
 // +kubebuilder:rbac:groups=core,resources=namespaces,verbs=get;list;watch
 // +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;create;update;patch;delete
 // +kubebuilder:rbac:groups=core,resources=secrets,verbs=get;list;create;update;patch;delete
+
 func (r *Reconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	switch r.resourceType {
 	case NamespaceType:
